Document the standard C library interface.

+++ ctangle.c
@@ -2,20 +2,17 @@
 #line 66 "ctangle.w"
 
 /*4:*/
-#line 44 "common.h"
-
-#include <stdio.h> 
-#include <stdbool.h> 
-#include <stdint.h> 
-#include <string.h> 
-
-/*:4*//*65:*/
-#line 864 "ctangle.w"
+#line 46 "common.h"
 
 #include <ctype.h>  
+#include <stdbool.h>  
+#include <stddef.h>  
+#include <stdint.h>  
 #include <stdlib.h>  
+#include <stdio.h>  
+#include <string.h>  
 
-/*:65*/
+/*:4*/
 #line 67 "ctangle.w"
 
 #define banner "This is CTANGLE (Version 3.65)" \
@@ -162,7 +159,7 @@ extern boolean program;
 extern int phase;
 
 /*:3*//*5:*/
-#line 69 "common.h"
+#line 74 "common.h"
 
 extern char section_text[];
 extern char*section_text_end;
@@ -170,7 +167,7 @@ extern char*id_first;
 extern char*id_loc;
 
 /*:5*//*6:*/
-#line 83 "common.h"
+#line 88 "common.h"
 
 extern char buffer[];
 extern char*buffer_end;
@@ -178,7 +175,7 @@ extern char*loc;
 extern char*limit;
 
 /*:6*//*7:*/
-#line 98 "common.h"
+#line 103 "common.h"
 
 typedef struct name_info{
 char*byte_start;
@@ -203,12 +200,12 @@ extern hash_pointer hash_end;
 extern hash_pointer h;
 
 /*:7*//*9:*/
-#line 143 "common.h"
+#line 148 "common.h"
 
 extern int history;
 
 /*:9*//*11:*/
-#line 163 "common.h"
+#line 168 "common.h"
 
 extern int include_depth;
 extern FILE*file[];
@@ -228,7 +225,7 @@ extern boolean changing;
 extern boolean web_file_open;
 
 /*:11*//*13:*/
-#line 187 "common.h"
+#line 192 "common.h"
 
 extern sixteen_bits section_count;
 extern boolean changed_section[];
@@ -236,14 +233,14 @@ extern boolean change_pending;
 extern boolean print_where;
 
 /*:13*//*14:*/
-#line 200 "common.h"
+#line 205 "common.h"
 
 extern int argc;
 extern char**argv;
 extern boolean flags[];
 
 /*:14*//*15:*/
-#line 212 "common.h"
+#line 217 "common.h"
 
 extern FILE*C_file;
 extern FILE*tex_file;
@@ -347,22 +344,22 @@ boolean comment_continues= 0;
 name_pointer cur_section_name;
 int no_where;
 
-/*:64*//*78:*/
-#line 1175 "ctangle.w"
+/*:64*//*77:*/
+#line 1171 "ctangle.w"
 
 text_pointer cur_text;
 eight_bits next_control;
 
-/*:78*//*85:*/
-#line 1333 "ctangle.w"
+/*:77*//*84:*/
+#line 1329 "ctangle.w"
 
 extern sixteen_bits section_count;
 
-/*:85*/
+/*:84*/
 #line 71 "ctangle.w"
 
 /*8:*/
-#line 121 "common.h"
+#line 126 "common.h"
 
 extern boolean names_match(name_pointer,const char*,size_t,eight_bits);
 extern name_pointer id_lookup(const char*,const char*,char);
@@ -376,7 +373,7 @@ extern void print_section_name(name_pointer);
 extern void sprint_section_name(char*,name_pointer);
 
 /*:8*//*10:*/
-#line 146 "common.h"
+#line 151 "common.h"
 
 extern int wrap_up(void);
 extern void err_print(const char*);
@@ -384,14 +381,14 @@ extern void fatal(const char*,const char*);
 extern void overflow(const char*);
 
 /*:10*//*12:*/
-#line 181 "common.h"
+#line 186 "common.h"
 
 extern boolean get_line(void);
 extern void check_complete(void);
 extern void reset_input(void);
 
 /*:12*//*16:*/
-#line 220 "common.h"
+#line 225 "common.h"
 
 extern void common_init(void);
 extern void print_stats(void);
@@ -411,18 +408,18 @@ static void output_defs(void);
 
 static void out_char(eight_bits);
 
-/*:51*//*93:*/
-#line 1440 "ctangle.w"
+/*:51*//*92:*/
+#line 1436 "ctangle.w"
 
 static void phase_one(void);
 
-/*:93*//*95:*/
-#line 1458 "ctangle.w"
+/*:92*//*94:*/
+#line 1454 "ctangle.w"
 
 static void skip_limbo(void);
 
-/*:95*//*99:*/
-#line 1533 "ctangle.w"
+/*:94*//*98:*/
+#line 1529 "ctangle.w"
 
 static eight_bits get_next(void);
 static eight_bits skip_ahead(void);
@@ -435,7 +432,7 @@ static void scan_repl(eight_bits);
 static void scan_section(void);
 static void store_two_bytes(sixteen_bits);
 
-/*:99*/
+/*:98*/
 #line 72 "ctangle.w"
 
 
@@ -497,11 +494,11 @@ ccode['<']= ccode['(']= section_name;
 ccode['\'']= ord;
 }
 
-/*:60*//*74:*/
-#line 1096 "ctangle.w"
+/*:60*//*73:*/
+#line 1092 "ctangle.w"
 section_text[0]= ' ';
 
-/*:74*/
+/*:73*/
 #line 88 "ctangle.w"
 
 common_init();
@@ -912,8 +909,8 @@ else loc++;
 }
 }
 
-/*:63*//*66:*/
-#line 876 "ctangle.w"
+/*:63*//*65:*/
+#line 872 "ctangle.w"
 
 static eight_bits
 get_next(void)
@@ -926,8 +923,8 @@ if(preprocessing&&*(limit-1)!='\\')preprocessing= 0;
 if(get_line()==0)return(new_section);
 else if(print_where&&!no_where){
 print_where= 0;
-/*80:*/
-#line 1205 "ctangle.w"
+/*79:*/
+#line 1201 "ctangle.w"
 
 store_two_bytes(0150000);
 if(changing&&include_depth==change_depth){
@@ -941,8 +938,8 @@ id_loc= id_first+strlen(id_first);
 {int a_l= id_lookup(id_first,id_loc,0)-name_dir;app_repl((a_l/0400)+0200);
 app_repl(a_l%0400);}
 
-/*:80*/
-#line 888 "ctangle.w"
+/*:79*/
+#line 884 "ctangle.w"
 
 }
 else return('\n');
@@ -955,8 +952,8 @@ if(comment_continues)return('\n');
 else continue;
 }
 loc++;
-if(xisdigit(c)||c=='.')/*69:*/
-#line 957 "ctangle.w"
+if(xisdigit(c)||c=='.')/*68:*/
+#line 953 "ctangle.w"
 {
 id_first= loc-1;
 if(*id_first=='.'&&!xisdigit(*loc))goto mistake;
@@ -980,14 +977,14 @@ id_loc= loc;
 return(constant);
 }
 
-/*:69*/
-#line 900 "ctangle.w"
+/*:68*/
+#line 896 "ctangle.w"
 
 else if(c=='\''||c=='"'
 ||((c=='L'||c=='u'||c=='U')&&(*loc=='\''||*loc=='"'))
 ||((c=='u'&&*loc=='8')&&(*(loc+1)=='\''||*(loc+1)=='"')))
-/*70:*/
-#line 985 "ctangle.w"
+/*69:*/
+#line 981 "ctangle.w"
 {
 char delim= c;
 id_first= section_text+1;
@@ -1030,12 +1027,12 @@ id_loc++;
 return(string);
 }
 
-/*:70*/
-#line 904 "ctangle.w"
+/*:69*/
+#line 900 "ctangle.w"
 
 else if(isalpha(c)||isxalpha(c)||ishigh(c))
-/*68:*/
-#line 950 "ctangle.w"
+/*67:*/
+#line 946 "ctangle.w"
 {
 id_first= --loc;
 while(isalpha((eight_bits)*++loc)||isdigit((eight_bits)*loc)
@@ -1043,11 +1040,11 @@ while(isalpha((eight_bits)*++loc)||isdigit((eight_bits)*loc)
 id_loc= loc;return(identifier);
 }
 
-/*:68*/
-#line 906 "ctangle.w"
+/*:67*/
+#line 902 "ctangle.w"
 
-else if(c=='@')/*71:*/
-#line 1030 "ctangle.w"
+else if(c=='@')/*70:*/
+#line 1026 "ctangle.w"
 {
 c= ccode[(eight_bits)*loc++];
 switch(c){
@@ -1062,12 +1059,12 @@ err_print("! Double @ should be used in control text");
 continue;
 case section_name:
 cur_section_name_char= *(loc-1);
-/*73:*/
-#line 1078 "ctangle.w"
+/*72:*/
+#line 1074 "ctangle.w"
 {
 char*k;
-/*75:*/
-#line 1098 "ctangle.w"
+/*74:*/
+#line 1094 "ctangle.w"
 
 k= section_text;
 while(1){
@@ -1077,8 +1074,8 @@ err_print("! Input ended in section name");
 loc= buffer+1;break;
 }
 c= *loc;
-/*76:*/
-#line 1122 "ctangle.w"
+/*75:*/
+#line 1118 "ctangle.w"
 
 if(c=='@'){
 c= *(loc+1);
@@ -1096,8 +1093,8 @@ err_print("! Nesting of section names not allowed");break;
 *(++k)= '@';loc++;
 }
 
-/*:76*/
-#line 1107 "ctangle.w"
+/*:75*/
+#line 1103 "ctangle.w"
 
 loc++;if(k<section_text_end)k++;
 if(xisspace(c)){
@@ -1113,8 +1110,8 @@ printf("...");mark_harmless;
 }
 if(*k==' '&&k> section_text)k--;
 
-/*:75*/
-#line 1080 "ctangle.w"
+/*:74*/
+#line 1076 "ctangle.w"
 
 if(k-section_text> 3&&strncmp(k-2,"...",3)==0)
 cur_section_name= section_lookup(section_text+1,k-3,1);
@@ -1137,16 +1134,16 @@ overflow("output files");
 }
 
 /*:43*/
-#line 1086 "ctangle.w"
+#line 1082 "ctangle.w"
 
 return(section_name);
 }
 
-/*:73*/
-#line 1044 "ctangle.w"
+/*:72*/
+#line 1040 "ctangle.w"
 
-case string:/*77:*/
-#line 1144 "ctangle.w"
+case string:/*76:*/
+#line 1140 "ctangle.w"
 {
 id_first= loc++;*(limit+1)= '@';*(limit+2)= '>';
 while(*loc!='@'||*(loc+1)!='>')loc++;
@@ -1156,11 +1153,11 @@ id_loc= loc;loc+= 2;
 return(string);
 }
 
-/*:77*/
-#line 1045 "ctangle.w"
+/*:76*/
+#line 1041 "ctangle.w"
 
-case ord:/*72:*/
-#line 1057 "ctangle.w"
+case ord:/*71:*/
+#line 1053 "ctangle.w"
 
 id_first= loc;
 if(*loc=='\\'){
@@ -1182,15 +1179,15 @@ err_print("! String didn't end");loc= limit-1;break;
 loc++;
 return(ord);
 
-/*:72*/
-#line 1046 "ctangle.w"
+/*:71*/
+#line 1042 "ctangle.w"
 
 default:return(c);
 }
 }
 
-/*:71*/
-#line 907 "ctangle.w"
+/*:70*/
+#line 903 "ctangle.w"
 
 else if(xisspace(c)){
 if(!preprocessing||loc> limit)continue;
@@ -1198,8 +1195,8 @@ if(!preprocessing||loc> limit)continue;
 else return(' ');
 }
 else if(c=='#'&&loc==buffer+1)preprocessing= 1;
-mistake:/*67:*/
-#line 928 "ctangle.w"
+mistake:/*66:*/
+#line 924 "ctangle.w"
 
 switch(c){
 case'+':if(*loc=='+')compress(plus_plus);break;
@@ -1222,23 +1219,23 @@ case'|':if(*loc=='|')compress(or_or);break;
 case'!':if(*loc=='=')compress(non_eq);break;
 }
 
-/*:67*/
-#line 914 "ctangle.w"
+/*:66*/
+#line 910 "ctangle.w"
 
 return(c);
 }
 }
 
-/*:66*//*79:*/
-#line 1179 "ctangle.w"
+/*:65*//*78:*/
+#line 1175 "ctangle.w"
 
 static void
 scan_repl(
 eight_bits t)
 {
 sixteen_bits a;
-if(t==section_name){/*80:*/
-#line 1205 "ctangle.w"
+if(t==section_name){/*79:*/
+#line 1201 "ctangle.w"
 
 store_two_bytes(0150000);
 if(changing&&include_depth==change_depth){
@@ -1252,20 +1249,20 @@ id_loc= id_first+strlen(id_first);
 {int a_l= id_lookup(id_first,id_loc,0)-name_dir;app_repl((a_l/0400)+0200);
 app_repl(a_l%0400);}
 
-/*:80*/
-#line 1185 "ctangle.w"
+/*:79*/
+#line 1181 "ctangle.w"
 }
 while(1)switch(a= get_next()){
-/*81:*/
-#line 1218 "ctangle.w"
+/*80:*/
+#line 1214 "ctangle.w"
 
 case identifier:a= id_lookup(id_first,id_loc,0)-name_dir;
 app_repl((a/0400)+0200);
 app_repl(a%0400);break;
 case section_name:if(t!=section_name)goto done;
 else{
-/*82:*/
-#line 1251 "ctangle.w"
+/*81:*/
+#line 1247 "ctangle.w"
 {
 char*try_loc= loc;
 while(*try_loc==' '&&try_loc<limit)try_loc++;
@@ -1277,14 +1274,14 @@ if(*try_loc=='=')err_print("! Missing `@ ' before a named section");
 
 }
 
-/*:82*/
-#line 1224 "ctangle.w"
+/*:81*/
+#line 1220 "ctangle.w"
 
 a= cur_section_name-name_dir;
 app_repl((a/0400)+0250);
 app_repl(a%0400);
-/*80:*/
-#line 1205 "ctangle.w"
+/*79:*/
+#line 1201 "ctangle.w"
 
 store_two_bytes(0150000);
 if(changing&&include_depth==change_depth){
@@ -1298,8 +1295,8 @@ id_loc= id_first+strlen(id_first);
 {int a_l= id_lookup(id_first,id_loc,0)-name_dir;app_repl((a_l/0400)+0200);
 app_repl(a_l%0400);}
 
-/*:80*/
-#line 1228 "ctangle.w"
+/*:79*/
+#line 1224 "ctangle.w"
 break;
 }
 case output_defs_code:if(t!=section_name)err_print("! Misplaced @h");
@@ -1309,8 +1306,8 @@ output_defs_seen= 1;
 a= output_defs_flag;
 app_repl((a/0400)+0200);
 app_repl(a%0400);
-/*80:*/
-#line 1205 "ctangle.w"
+/*79:*/
+#line 1201 "ctangle.w"
 
 store_two_bytes(0150000);
 if(changing&&include_depth==change_depth){
@@ -1324,14 +1321,14 @@ id_loc= id_first+strlen(id_first);
 {int a_l= id_lookup(id_first,id_loc,0)-name_dir;app_repl((a_l/0400)+0200);
 app_repl(a_l%0400);}
 
-/*:80*/
-#line 1237 "ctangle.w"
+/*:79*/
+#line 1233 "ctangle.w"
 
 }
 break;
 case constant:case string:
-/*83:*/
-#line 1262 "ctangle.w"
+/*82:*/
+#line 1258 "ctangle.w"
 
 app_repl(a);
 while(id_first<id_loc){
@@ -1344,12 +1341,12 @@ app_repl(*id_first++);
 }
 app_repl(a);break;
 
-/*:83*/
-#line 1241 "ctangle.w"
+/*:82*/
+#line 1237 "ctangle.w"
 
 case ord:
-/*84:*/
-#line 1278 "ctangle.w"
+/*83:*/
+#line 1274 "ctangle.w"
 {
 int c= (eight_bits)*id_first;
 if(c=='\\'){
@@ -1399,8 +1396,8 @@ app_repl(constant);
 }
 break;
 
-/*:84*/
-#line 1243 "ctangle.w"
+/*:83*/
+#line 1239 "ctangle.w"
 
 case definition:case format_code:case begin_C:if(t!=section_name)goto done;
 else{
@@ -1409,8 +1406,8 @@ err_print("! @d, @f and @c are ignored in C text");continue;
 }
 case new_section:goto done;
 
-/*:81*/
-#line 1190 "ctangle.w"
+/*:80*/
+#line 1186 "ctangle.w"
 
 case')':app_repl(a);
 if(t==macro)app_repl(' ');
@@ -1422,8 +1419,8 @@ if(text_ptr> text_info_end)overflow("text");
 cur_text= text_ptr;(++text_ptr)->tok_start= tok_ptr;
 }
 
-/*:79*//*86:*/
-#line 1340 "ctangle.w"
+/*:78*//*85:*/
+#line 1336 "ctangle.w"
 
 static void
 scan_section(void)
@@ -1437,8 +1434,8 @@ printf("*%d",section_count);update_terminal;
 }
 next_control= 0;
 while(1){
-/*87:*/
-#line 1379 "ctangle.w"
+/*86:*/
+#line 1375 "ctangle.w"
 
 while(next_control<definition)
 
@@ -1446,12 +1443,12 @@ if((next_control= skip_ahead())==section_name){
 loc-= 2;next_control= get_next();
 }
 
-/*:87*/
-#line 1354 "ctangle.w"
+/*:86*/
+#line 1350 "ctangle.w"
 
 if(next_control==definition){
-/*88:*/
-#line 1386 "ctangle.w"
+/*87:*/
+#line 1382 "ctangle.w"
 {
 while((next_control= get_next())=='\n');
 if(next_control!=identifier){
@@ -1469,8 +1466,8 @@ scan_repl(macro);
 cur_text->text_link= 0;
 }
 
-/*:88*/
-#line 1356 "ctangle.w"
+/*:87*/
+#line 1352 "ctangle.w"
 
 continue;
 }
@@ -1479,36 +1476,36 @@ p= name_dir;break;
 }
 if(next_control==section_name){
 p= cur_section_name;
-/*89:*/
-#line 1411 "ctangle.w"
+/*88:*/
+#line 1407 "ctangle.w"
 
 while((next_control= get_next())=='+');
 if(next_control!='='&&next_control!=eq_eq)
 continue;
 
-/*:89*/
-#line 1364 "ctangle.w"
+/*:88*/
+#line 1360 "ctangle.w"
 
 break;
 }
 return;
 }
 no_where= print_where= 0;
-/*90:*/
-#line 1416 "ctangle.w"
+/*89:*/
+#line 1412 "ctangle.w"
 
-/*91:*/
-#line 1421 "ctangle.w"
+/*90:*/
+#line 1417 "ctangle.w"
 
 store_two_bytes((sixteen_bits)(0150000+section_count));
 
 
-/*:91*/
-#line 1417 "ctangle.w"
+/*:90*/
+#line 1413 "ctangle.w"
 
 scan_repl(section_name);
-/*92:*/
-#line 1425 "ctangle.w"
+/*91:*/
+#line 1421 "ctangle.w"
 
 if(p==name_dir||p==0){
 (last_unnamed)->text_link= cur_text-text_info;last_unnamed= cur_text;
@@ -1524,17 +1521,17 @@ q->text_link= cur_text-text_info;
 cur_text->text_link= section_flag;
 
 
-/*:92*/
-#line 1419 "ctangle.w"
+/*:91*/
+#line 1415 "ctangle.w"
 
 
-/*:90*/
-#line 1370 "ctangle.w"
+/*:89*/
+#line 1366 "ctangle.w"
 
 }
 
-/*:86*//*94:*/
-#line 1443 "ctangle.w"
+/*:85*//*93:*/
+#line 1439 "ctangle.w"
 
 static void
 phase_one(void){
@@ -1547,8 +1544,8 @@ check_complete();
 phase= 2;
 }
 
-/*:94*//*96:*/
-#line 1461 "ctangle.w"
+/*:93*//*95:*/
+#line 1457 "ctangle.w"
 
 static void
 skip_limbo(void)
@@ -1562,8 +1559,8 @@ if(loc++<=limit){
 c= *loc++;
 if(ccode[(eight_bits)c]==new_section)break;
 switch(ccode[(eight_bits)c]){
-case translit_code:/*97:*/
-#line 1490 "ctangle.w"
+case translit_code:/*96:*/
+#line 1486 "ctangle.w"
 
 while(xisspace(*loc)&&loc<limit)loc++;
 loc+= 3;
@@ -1587,8 +1584,8 @@ translit[i-0200][loc-beg]= '\0';
 }
 }
 
-/*:97*/
-#line 1474 "ctangle.w"
+/*:96*/
+#line 1470 "ctangle.w"
 break;
 case format_code:case'@':break;
 case control_text:if(c=='q'||c=='Q'){
@@ -1605,20 +1602,20 @@ default:err_print("! Double @ should be used in limbo");
 }
 }
 
-/*:96*//*98:*/
-#line 1516 "ctangle.w"
+/*:95*//*97:*/
+#line 1512 "ctangle.w"
 
 void
 print_stats(void){
 puts("\nMemory usage statistics:");
 printf("%ld names (out of %ld)\n",
-(long)(name_ptr-name_dir),(long)max_names);
+(ptrdiff_t)(name_ptr-name_dir),(long)max_names);
 printf("%ld replacement texts (out of %ld)\n",
-(long)(text_ptr-text_info),(long)max_texts);
+(ptrdiff_t)(text_ptr-text_info),(long)max_texts);
 printf("%ld bytes (out of %ld)\n",
-(long)(byte_ptr-byte_mem),(long)max_bytes);
+(ptrdiff_t)(byte_ptr-byte_mem),(long)max_bytes);
 printf("%ld tokens (out of %ld)\n",
-(long)(tok_ptr-tok_mem),(long)max_toks);
+(ptrdiff_t)(tok_ptr-tok_mem),(long)max_toks);
 }
 
-/*:98*/
+/*:97*/
